refactor(themes): structure files and add dark themes
