+++ src/components/button/icon-button.ts
@@ -2,7 +2,7 @@ import { html } from 'lit';
 import { property } from 'lit/decorators.js';
 import { ifDefined } from 'lit/directives/if-defined.js';
 import { IgcButtonBaseComponent } from './button-base.js';
-import { styles } from './icon-button.fluent.css';
+import { styles } from './themes/icon-button/light/icon-button.bootstrap.css';
 import '../icon/icon';
 import { alternateName, blazorInclude } from '../common/decorators';
 import {
