+++ src/components/button/themes/icon-button/light/icon-button.base.scss
@@ -1,5 +1,5 @@
-@use '../../styles/common/component';
-@use '../../styles/utilities' as *;
+@use '../../../../../styles/common/component';
+@use '../../../../../styles/utilities' as *;
 
 :host {
     display: inline-flex;
