+++ src/components/button/themes/icon-button/light/icon-button.bootstrap.scss
@@ -1,5 +1,5 @@
 @use './icon-button.base.scss';
-@use '../../styles/utilities' as utils;
+@use '../../../../../styles/utilities' as utils;
 
 button {
     border-radius: utils.rem(4px);
