+++ src/components/button/themes/icon-button/light/icon-button.fluent.scss
@@ -1,5 +1,5 @@
 @use './icon-button.base.scss';
-@use '../../styles/utilities' as utils;
+@use '../../../../../styles/utilities' as utils;
 
 $button-color: #{utils.contrast-color(primary, 600)} !default;
 $hover-color: #{utils.color(primary, 600)} !default;
