--- src/components/list/list.material.scss
@@ -1,1 +1,0 @@
-@use './list.base.scss';
