+++ src/components/list/themes/light/list.indigo.scss
@@ -1,3 +1,5 @@
+@use '../../../../styles/common/component';
+@use '../../../../styles/utilities' as *;
 @use './list.base.scss';
 
 $header-color: color(primary, 500) !default;
