+++ src/components/list/themes/light/list.material.scss
@@ -1,0 +1,3 @@
+@use '../../../../styles/common/component';
+@use '../../../../styles/utilities' as *;
+@use './list.base.scss';
